Fixes #147
Convert empty state layout to enum
