+++ Example/Podfile
@@ -15,14 +15,14 @@ end
 
 
 
-# post_install do |installer|
-# 	myTargets = ['Nimble', 'Quick']
-#
-# 	installer.pods_project.targets.each do |target|
-# 		if myTargets.include? target.name
-# 			target.build_configurations.each do |config|
-# 				config.build_settings['SWIFT_VERSION'] = '3.2'
-# 			end
-# 		end
-# 	end
-# end
+post_install do |installer|
+	myTargets = ['Nimble', 'Quick']
+
+	installer.pods_project.targets.each do |target|
+		if myTargets.include? target.name
+			target.build_configurations.each do |config|
+				config.build_settings['SWIFT_VERSION'] = '3.2'
+			end
+		end
+	end
+end
