+++ QminderAPI/Classes/Models/EmptyState.swift
@@ -8,6 +8,30 @@
 
 import Foundation
 
+/// Empty state layout
+public enum EmptyStateLayout: String, Codable {
+  
+  /// Closed
+  case closed = "CLOSED"
+  
+  /// Simple
+  case simple = "SIMPLE"
+  
+  /// Other (not specified)
+  case other
+  
+  public init?(rawValue: String) {
+    switch rawValue.lowercased() {
+    case "simple":
+      self = .simple
+    case "closed":
+      self = .closed
+    default:
+      self = .other
+    }
+  }
+}
+
 /// Empty state object
 public struct EmptyState: CodableResponsable {
   
@@ -15,7 +39,7 @@ public struct EmptyState: CodableResponsable {
   let statusCode: Int
   
   /// Empty state layout
-  public let layout: String
+  public let layout: EmptyStateLayout
   
   /// Empty state message
   public let message: String
